star-ops-ux: Spawn view enhancements
- refactor using Indigo components
- shuffle suggested point names
- click point to populate name field
- click Address button to populate ownership field
